Update emcee_lowE_reco.py
The prior initialization of the walkers in time was well motivated - we know that the average cluster hit time is (from MC) typically ~10ns after the emission time for low energy electrons. However, where the algorithm placed the initial positions of the walkers in time was actually offset (not centered around the truth emission time). This is because the code was calculating the average hit time of the filtered hits, whose times are much earlier than the average cluster time. This commit changes it to the average cluster time instead. Also can add a larger spread (if needed) to the initial times. Right now it's set to a 2ns spread.